Self-parsing sections
* Each section parses itself
* Remove redundant code
* Add parsing utility methods for common types on those types
