+++ wasm/src/main/java/com/dylibso/chicory/wasm/types/DataSection.java
@@ -1,6 +1,11 @@
 package com.dylibso.chicory.wasm.types;
 
+import com.dylibso.chicory.wasm.Parser;
+import com.dylibso.chicory.wasm.exceptions.ChicoryException;
+import com.dylibso.chicory.wasm.io.WasmIOException;
+import com.dylibso.chicory.wasm.io.WasmInputStream;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 public final class DataSection extends Section {
@@ -50,5 +55,29 @@ public final class DataSection extends Section {
         int idx = dataSegments.size();
         dataSegments.add(dataSegment);
         return idx;
+    }
+
+    public void readFrom(final WasmInputStream in) throws WasmIOException {
+        var dataSegmentCount = in.u31();
+        dataSegments.ensureCapacity(dataSegments.size() + dataSegmentCount);
+
+        for (var i = 0; i < dataSegmentCount; i++) {
+            var mode = in.u32();
+            if (mode == 0) {
+                var offset = Parser.parseExpression(in);
+                byte[] data = in.byteVec();
+                addDataSegment(new ActiveDataSegment(List.of(offset), data));
+            } else if (mode == 1) {
+                byte[] data = in.byteVec();
+                addDataSegment(new PassiveDataSegment(data));
+            } else if (mode == 2) {
+                var memoryId = in.u31();
+                var offset = Parser.parseExpression(in);
+                byte[] data = in.byteVec();
+                addDataSegment(new ActiveDataSegment(memoryId, List.of(offset), data));
+            } else {
+                throw new ChicoryException("Failed to parse data segment with data mode: " + mode);
+            }
+        }
     }
 }
