+++ wasm/src/main/java/com/dylibso/chicory/wasm/types/ImportSection.java
@@ -1,5 +1,8 @@
 package com.dylibso.chicory.wasm.types;
 
+import com.dylibso.chicory.wasm.io.WasmIOException;
+import com.dylibso.chicory.wasm.io.WasmInputStream;
+import com.dylibso.chicory.wasm.io.WasmParseException;
 import java.util.ArrayList;
 import java.util.Objects;
 import java.util.stream.Stream;
@@ -51,5 +54,49 @@ public final class ImportSection extends Section {
         int idx = imports.size();
         imports.add(import_);
         return idx;
+    }
+
+    public void readFrom(final WasmInputStream in) throws WasmIOException {
+        var importCount = in.u31();
+        imports.ensureCapacity(imports.size() + importCount);
+
+        // Parse individual imports in the import section
+        for (int i = 0; i < importCount; i++) {
+            String moduleName = in.utf8();
+            String importName = in.utf8();
+            var descType = ExternalType.byId(in.u8());
+            switch (descType) {
+                case FUNCTION:
+                    {
+                        addImport(new FunctionImport(moduleName, importName, in.u31()));
+                        break;
+                    }
+                case TABLE:
+                    {
+                        var tableType = in.refType();
+                        var limitType = in.u32();
+                        if (limitType != 0x00 && limitType != 0x01) {
+                            throw new WasmParseException("Invalid limit type");
+                        }
+                        var min = in.u32Long();
+                        var limits =
+                                limitType > 0 ? new Limits(min, in.u32Long()) : new Limits(min);
+
+                        addImport(new TableImport(moduleName, importName, tableType, limits));
+                        break;
+                    }
+                case MEMORY:
+                    {
+                        MemoryLimits limits = MemoryLimits.parseFrom(in);
+                        addImport(new MemoryImport(moduleName, importName, limits));
+                        break;
+                    }
+                case GLOBAL:
+                    var globalValType = in.type();
+                    var globalMut = in.mut();
+                    addImport(new GlobalImport(moduleName, importName, globalMut, globalValType));
+                    break;
+            }
+        }
     }
 }
