+++ wasm/src/main/java/com/dylibso/chicory/wasm/types/Section.java
@@ -1,5 +1,8 @@
 package com.dylibso.chicory.wasm.types;
 
+import com.dylibso.chicory.wasm.io.WasmIOException;
+import com.dylibso.chicory.wasm.io.WasmInputStream;
+
 /**
  * A section of a WASM file.
  */
@@ -21,4 +24,13 @@ public abstract class Section {
     public int sectionId() {
         return id;
     }
+
+    /**
+     * Read this section from the given input.
+     * Any bytes in this section not consumed will be skipped after this method returns.
+     *
+     * @param in the input (not {@code null})
+     * @throws WasmIOException if an I/O error occurs while reading the section
+     */
+    public abstract void readFrom(WasmInputStream in) throws WasmIOException;
 }
