+++ wasm/src/main/java/com/dylibso/chicory/wasm/types/UnknownCustomSection.java
@@ -1,5 +1,7 @@
 package com.dylibso.chicory.wasm.types;
 
+import com.dylibso.chicory.wasm.io.WasmIOException;
+import com.dylibso.chicory.wasm.io.WasmInputStream;
 import java.util.Objects;
 
 /**
@@ -19,5 +21,9 @@ public final class UnknownCustomSection extends CustomSection {
 
     public String name() {
         return name;
+    }
+
+    public void readFrom(final WasmInputStream in) throws WasmIOException {
+        // do nothing
     }
 }
