+++ wasm/src/main/java/com/dylibso/chicory/wasm/types/UnknownSection.java
@@ -1,5 +1,8 @@
 package com.dylibso.chicory.wasm.types;
 
+import com.dylibso.chicory.wasm.io.WasmIOException;
+import com.dylibso.chicory.wasm.io.WasmInputStream;
+
 /**
  * A section which is not known to the parser.
  */
@@ -11,5 +14,9 @@ public final class UnknownSection extends Section {
      */
     public UnknownSection(final int id) {
         super(id);
+    }
+
+    public void readFrom(final WasmInputStream in) throws WasmIOException {
+        // do nothing
     }
 }
